Profile page basics showing. Routing working for live map, vehicle profile, alerts, and routing page.
