+++ client/pages/vehicle-list.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
-
+import VehicleProfile from './vehicle-profile';
 export default class VehicleList extends React.Component {
 
   render() {
 
     return (
-     <div>
-       <p>somelist </p>
-     </div>
+
+       <VehicleProfile />
+
     );
   }
 }
